Remove texmode hacks, use standard Unicode input
I can't debug stuff that can't be compiled using standard SILE because
the trace functions are useless when using input filters. And honestly
it isn't needed here. For a handful of TeX muscle memory shortcuts you
end up with a document that can't easily be compiled.

This is a Unicode proposal. I don't think using Unicode input is much of
a stretch.